Rename StickyNav to StickyHeader
so as to not create false assumptions about its functionality.
