+++ ts/change-notetype/ChangeNotetypePage.svelte
@@ -10,7 +10,7 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
     import Row from "../components/Row.svelte";
     import Col from "../components/Col.svelte";
     import NotetypeSelector from "./NotetypeSelector.svelte";
-    import StickyNav from "./StickyNav.svelte";
+    import StickyHeader from "./StickyHeader.svelte";
     import Mapper from "./Mapper.svelte";
 
     export let state: ChangeNotetypeState;
@@ -26,13 +26,13 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
     <Row class="gx-0" --cols={2}>
         <Col --col-size={1} breakpoint="md">
             <Container>
-                <StickyNav {state} ctx={MapContext.Field} />
+                <StickyHeader {state} ctx={MapContext.Field} />
                 <Mapper {state} ctx={MapContext.Field} />
             </Container>
         </Col>
         <Col --col-size={1} breakpoint="md">
             <Container>
-                <StickyNav {state} ctx={MapContext.Template} />
+                <StickyHeader {state} ctx={MapContext.Template} />
                 {#if $info.templates}
                     <Mapper {state} ctx={MapContext.Template} />
                 {:else}
